+++ lionengine-core/src/main/java/com/b3dgs/lionengine/geom/Polygon.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2013-2014 Byron 3D Games Studio (www.b3dgs.com) Pierre-Alexandre (contact@b3dgs.com)
+ * Copyright (C) 2013-2015 Byron 3D Games Studio (www.b3dgs.com) Pierre-Alexandre (contact@b3dgs.com)
  * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -32,7 +32,7 @@ public interface Polygon
      * @param x The horizontal location.
      * @param y The vertical location.
      */
-    void addPoint(int x, int y);
+    void addPoint(double x, double y);
 
     /**
      * Reset the polygon.
